move python module app.db -> app.data.db
